requirements.txt
langchain>=0.1.20
langchain-community>=0.1.0
langgraph>=0.1.0
openai>=1.0.0
termcolor>=2.3.0
prompt_toolkit>=3.0.0
requests>=2.0.0
tzlocal
pydantic>=2.0.0
langchain_experimental
pyyaml>=6.0.0
langchain-tavily>=0.0.1
# Placeholder for MCP client, update as needed when actual MCP client is selected or implemented